Extract security exception classes to library
