+++ common-libraries/common-exception-handling/src/main/java/by/kovzov/uis/common/exception/handler/security/AuthenticationExceptionHandler.java
@@ -1,8 +1,5 @@
-package by.kovzov.uis.security.rest.security.exception;
+package by.kovzov.uis.common.exception.handler.security;
 
-import java.io.IOException;
-
-import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -13,15 +10,14 @@ import org.springframework.web.servlet.HandlerExceptionResolver;
 
 @Component
 @RequiredArgsConstructor
-public class ExceptionAuthenticationEntryPoint implements AuthenticationEntryPoint {
+public class AuthenticationExceptionHandler implements AuthenticationEntryPoint {
 
     @Qualifier("handlerExceptionResolver")
     private final HandlerExceptionResolver resolver;
 
     @Override
     public void commence(HttpServletRequest request, HttpServletResponse response,
-                         org.springframework.security.core.AuthenticationException authException)
-        throws IOException, ServletException {
+                         org.springframework.security.core.AuthenticationException authException) {
         resolver.resolveException(request, response, null, authException);
     }
 }
